+++ R/installed-packages.R
@@ -1,5 +1,5 @@
 
-installed_packages <- function(lib.loc = NULL, priority = NULL, ...) {
-  packages <- installed.packages(lib.loc = lib.loc, priority = priority, ...)
+installed_packages <- function(lib.loc = NULL, priority = NULL) {
+  packages <- installed.packages(lib.loc = lib.loc, priority = priority)
   as_data_frame(packages)
 }