+++ R/packages.R
@@ -4,7 +4,7 @@
 renv_packages_base <- function() {
 
   `_renv_packages_base` <<- `_renv_packages_base` %||% {
-    db <- installed_packages(library = .Library, priority = "base")
+    db <- installed_packages(lib.loc = .Library, priority = "base")
     c("R", db$Package, "translations")
   }